# [fpga,rom] Speed up ROM splicing by omitting contiguous addresses