Although the internal operations of UpdateMEM are a black box, we can surmise from its documentation that it likely has a complexity related to the number of both the data and address values in the generated MEM file, though it is not clear what dominates runtime. The [UpdateMEM documentation](https://docs.amd.com/r/2024.1-English/ug1580-updatemem/Bit-Lane-Definitions-Memory-Device-Usage) does suggest however that UpdateMEM will check for violations in the MEM file, which will scale with address count.

Currently, when generating an UpdateMEM-compatible MEM file for the ROM, we emit an address for each individual ROM word (unlike the OTP, which is one contiguous block). This is despite the fact that, for the majority of widths, the number of characters per word will be such that data is written to contiguous addresses. For example:
```as
@00000000 1F5453B2E1
@00000005 6B008E0724
@0000000A B2684DC8B7
@0000000F C863063B45
...
```
which can be more succinctly represented as:
```as
@00000000 1F5453B2E1 6B008E0724 B2684DC8B7 C863063B45 ...
```
The [UpdateMEM documentation](https://docs.amd.com/r/2024.1-English/ug1580-updatemem/Memory-Files) even suggests that "UpdateMEM requires a less redundant format ... an address specifier is used only once at the beginning of a contiguous block of data." I've checked that this documentation is consistent for at least UpdateMEM 2021.2 through 2024.1.

This PR Introduces logic to check whether ROM word addresses are contiguous, and, if so, omit the redundant addresses from the generated MEM file.

*Aside*: I believe that the calculations for characters per word (`word_chars = math.ceil(width / 4)`) and addresses (`addr = idx * math.ceil(width / 8)` mean that generated words are either contiguous, or off by at most one nibble. However, per the [UpdateMEM documentation](https://docs.amd.com/r/2024.1-English/ug1580-updatemem/Bit-Lane-Definitions-Memory-Device-Usage): "when a value has an odd number of hexadecimal characters, the first hexadecimal character is assumed to be a zero" (e.g. `C74` -> `0C74`). So, even if e.g. `width = 25` such that `word_chars = 7` whilst the `addr` stride is 4, the prefixing by 0 means that writing e.g. `@00000000 0123456 789ABCD` should be read the same as `@00000000 0123456 @00000004 789ABCD`. All to say, I think that this condition does not need to be checked at all, and we could always do this, but I have opted to include it regardless for safety.

Building tests, it appears that UpdateMEM will then produce identical spliced bitstreams, whilst basic empirical testing (building `fpga_cw310_sival_rom_ext`  for the UART smoketest) suggests that the ROM splicing action speeds up by anywhere from 4-5x, though I suspect a decent part of the remaining time is overhead from Vivado & loading the bitstream:
- **Before this change**: average UpdateMEM CPU time: 52s, elapsed time: 54s
- **After this change**: average UpdateMEM CPU time: 11s, elapsed time: 11s

As ROM splicing lies on the critical path, this should help somewhat speed up cold builds, as well as bitstream & test builds following any HW/ROM changes, at the cost of no additional complexity. I'm not as familiar with FPGAs compared to many of the knowledgeable people working on OpenTitan, so please do let me know if there's any problems that I've missed with this approach.